Handle level filter parse errors and split logger tests
